pyproject.toml
[project]
name = "dify-chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daphne>=4.2.1",
    "dj-database-url>=3.1.0",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]
